backend/requirements.txt
asgiref==3.6.0
coverage==7.2.1
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-gis==1.0.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
